restore date formats and stop blocking call to google

+++ src/main/java/com/ociweb/schema/MessageScheme.java
@@ -93,8 +93,8 @@ public class MessageScheme {
             string,
             string,
             string,
-            integer,
-            integer,
+            int64,
+            int64,
     };
 
     public static TrieParser buildParser() {
